Revert creation of wgpu.gui.auto.qt
Also removed the size and title arguments for the QWgpuWidget
as it seems that they were a leftover from when the QWgpuWidget
was the top-level widget. Note that QWgpuCanvas still has those
arguments, being a proper top-level Qt widget.
